refactor(chat): extract chat_manager/deps_recovery.rs (#292)
Move RecoveryWrapper, its TranscriptPresenceDeps/DegradedRecoveryDeps
impls, and the recovery_wrapper accessor into their own file so the
wrapper's five fields stay private while the type itself is pub(super)
for the sibling group's later history.rs extraction.
